+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 - Migrate to TypeScript
 - Accept `site` config when instantiating the client
-- Export a `createClient` function (instead of a default value)
+- **Breaking change**: Export a `createClient` function (instead of a default value)
+- Add `invoke` method to the client
 
 ## 1.5.0